--- src/reviews-cont/filter/filter-type.js
@@ -1,11 +1,0 @@
-'use strict';
-
-var Filter = {
-  'ALL': 'reviews-all',
-  'RECENT': 'reviews-recent',
-  'GOOD': 'reviews-good',
-  'BAD': 'reviews-bad',
-  'POPULAR': 'reviews-popular'
-};
-
-module.exports = Filter;
